Re #626: Copy/move/import/export segment GUI and some of the logic; fixed minor bugs
Displayable manager does not call Modified on the representation before setting it as input to the pipeline as it triggers deletion of all other representations if it's the master
Show/hide segment hides the segment from the merged labelmap by setting the opacity for the color in the color table to zero. It does not trigger re-generation of merged labelmap, but if it would be desirable then it can be done later.