Re #760: Add function to disable representation. Useful for testing if one wants to use a non-default conversion path without manual UI interaction.
Example usage from python (before loading the dataset):
import vtkSegmentationCore
vtkSegmentationCore.vtkSegmentationConverterFactory.GetInstance().DisableRepresentation('Closed surface')

Now it is used to make sure the labelmaps are created via ribbon model, to verify that the test baselines match exactly with the old contour ones. Later when direct closed surface conversion is proven to be correct, the tests can be changed to use that.