re #774: Added a flag to stop the Matlab process after completion (based on Alexis Girault's patch)
If no command is specified then Matlab is not launched, therefore it is safe to run --exitmatlab with an empty command if the goal is to shutdown Matlab.
The difference compared to sending an "exit" command to Matlab is that if Matlab was not running and an exit command is sent then Matlab is started and then shut down.
Running Matlab Commander with an empty command and --exitmatlab shuts Matlab down if it was running and does not do anything if Matlab was not running.